+++ src/RaylibImpl/RaylibSprite.h
@@ -6,8 +6,8 @@ class RaylibSprite : public ISprite
 {
 public:
 	void LoadImage(const char* path) override;
-	void SetPosition(int x, int y) override;
-	void SetSize(int w, int h) override;
+	void SetPosition(float x, float y) override;
+	void SetSize(float w, float h) override;
 
 	Texture2D GetTexture() { return m_Text; }
